Test: move database tests out of util into their relevant modules
The util test package held tests that exercised other modules. Redistribute them:

- CalculationMetaData / _DatabaseData structural tests -> tests/raw/test_data.py
- _DatabaseModel guarantees (no schema_version, VaspData unwrap) ->
  tests/raw/test_model_validation.py
- Calculation._to_database() container/property-structure tests ->
  tests/calculation/test_default_calculation.py

Delete tests/util/test_to_database_new.py and drop the same-category strays from
tests/util/test_database.py, which now tests only py4vasp._util.database helpers.

Also scope schema_fingerprint() to _DatabaseModel subclasses defined in the
models module, so throwaway subclasses created in tests no longer leak into the
fingerprint (previously an ordering-dependent flake in the snapshot test).
